Rewrite address
Now I'm using the normal URL parser, with a few differences.
PeerID is in the username field, the token is in the password field.  The host is the same.

This has the added benefit of giving us a path that we could use later for something.  I'm not sure what yet, but it's there.